airflow/requirements.txt
yfinance>=0.2.40
psycopg2-binary>=2.9.9
redis>=5.0.8
